refactor(markdown): record conformance as a ratio against a named version
The baseline was a bare integer in a text file, and a bare integer says
nothing. `652` is one claim against CommonMark 0.31.2, which has 652 examples,
and quite another against a suite with more — and nothing in the file said
which suite, or which version of it, the number was ever measured from.

It is a JSON array now, one entry per profile, each carrying the profile, the
version, the fixture file, and both halves of the ratio. `total` is checked
against the fixture file rather than derived from it, which is the point:
vendoring a newer spec release now fails loudly and asks for the version, the
total and the passing count to move together, instead of quietly re-basing what
we claim.

An array because a second profile is foreseeable: a GFM suite has no CommonMark
fixtures to borrow and would want its own entry beside this one. The harness
reads the list and measures each, so adding one is appending an entry and
vendoring its fixtures — which is why it is no longer named after CommonMark.

Both guards are exercised: a stale ratio prints what to raise it to, an
impossible one fails as a regression, and a total that disagrees with the file
names both numbers and says what to do about it.

652/652 either way; no behaviour change.
