+++ morphir/langkit/markdown/scalatags/test/resources/conformance-baselines.json
@@ -1,0 +1,9 @@
+[
+  {
+    "profile": "CommonMark",
+    "version": "0.31.2",
+    "fixtures": "commonmark-0.31.2-spec.json",
+    "passing": 652,
+    "total": 652
+  }
+]
